requirements.txt
GitPython
signal
sys